requirements.txt
google-play-scraper
google-genai
python-dotenv
pandas
flask
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
